+++ Server/src/GameWorld/WorldObjects/Weapon/WeaponShotgun.cpp
@@ -19,8 +19,8 @@ void WeaponShotgun::attack(AttackInformation& attack_info,
 	if (activated && ammo >= 1) {
 		attack_info.set_weapon(PositionType::M3);
 		std::map<char,Character>::iterator closest_char;
-		float distance;
-		float angle = attack_info.get_angle();
+		double distance;
+		int angle = attack_info.get_angle();
 		bool is_character = find_closest_character(attack_info, blocks,
 								characters, angle, closest_char, distance);
 		if (is_character) {
@@ -41,7 +41,7 @@ char WeaponShotgun::get_type(){
 }
 
 
-char WeaponShotgun::calculate_damage(float distance) {
+char WeaponShotgun::calculate_damage(double distance) {
 	std::mt19937_64 rng;
 	uint64_t seed = std::chrono::high_resolution_clock::now().
 										time_since_epoch().count();
