+++ Server/src/GameWorld/WorldObjects/Weapon/WeaponWhite.h
@@ -15,7 +15,7 @@ class WeaponWhite : public Weapon {
 
         char get_type();
         //void shoot(Character &character, uint16_t distance);
-        char calculate_damage(float distance);
+        char calculate_damage(double distance);
 
         ~WeaponWhite();
 };
